prep_chroot_env.py : Separate the chroot creation into another script
Chroot creation is now a dedicated command-line invokable script.
This removes the need for ever having to use root to run the package creation
If, and only if, the chroot does not already exist will root be necessary.
It will also only be necessary for THIS script, and nothing else, which is
much safer and prevents everything created to be owned by root which is anoying.

Signed-off-by: Simon Beaudoin <sbeaudoi@qti.qualcomm.com>