Cleaner merge commit when merging new upstream
Using gbp-import-orig doesn't result in the correct commit graph. We
want a single merge commit where:

1. The tree takes the previous Debian packaging from the debian/
directory, the previous Debian-packaging-specific GitHub actions in
.github/, but everything else from the new upstream.

2. The first parent is the previous Debian packaging.

3. The second parent is the new upstream commit.

gbp-import-ref doesn't exactly do this either. --merge-mode=replace is
close, but it doesn't accommodate .github/. This seems like it would be
useful to get support for in upstream gbp-import-ref, which I intend to
request next.

In the meantime, we can arrange to do this ourselves wiht a fairly
straightforward script.

Signed-off-by: Robie Basak <robie.basak@oss.qualcomm.com>